LayoutTests/fast/dom/HTMLTemplateElement/cycles-expected.txt

Test that cycles are not allowed in template content

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS template.content.appendChild(template) threw exception HierarchyRequestError (DOM Exception 3): The operation would yield an incorrect node tree..
PASS template.content.appendChild(outerDiv) threw exception HierarchyRequestError (DOM Exception 3): The operation would yield an incorrect node tree..
PASS innerDiv.appendChild(template) threw exception HierarchyRequestError (DOM Exception 3): The operation would yield an incorrect node tree..
PASS innerDiv.appendChild(outerDiv) threw exception HierarchyRequestError (DOM Exception 3): The operation would yield an incorrect node tree..
PASS innerTemplate.appendChild(outerDiv) threw exception HierarchyRequestError (DOM Exception 3): The operation would yield an incorrect node tree..
PASS successfullyParsed is true

TEST COMPLETE
